+++ src/main/java/io/dataspaceconnector/common/ids/message/ClearingHouseService.java
@@ -15,19 +15,29 @@
  */
 package io.dataspaceconnector.common.ids.message;
 
+import java.net.URI;
+import java.util.Arrays;
+import java.util.UUID;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import de.fraunhofer.iais.eis.ContractAgreement;
 import de.fraunhofer.iais.eis.Message;
+import io.dataspaceconnector.common.exception.MessageResponseException;
 import io.dataspaceconnector.common.exception.PolicyExecutionException;
 import io.dataspaceconnector.common.exception.UUIDFormatException;
 import io.dataspaceconnector.common.util.UUIDUtils;
 import io.dataspaceconnector.config.ConnectorConfig;
+import io.dataspaceconnector.model.message.ProcessCreationMessageDesc;
 import io.dataspaceconnector.service.message.builder.type.LogMessageService;
+import io.dataspaceconnector.service.message.builder.type.ProcessCreationRequestService;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import net.minidev.json.JSONObject;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.util.UriComponentsBuilder;
-
-import java.net.URI;
 
 /**
  * Service for communication with the clearing house.
@@ -36,6 +46,18 @@ import java.net.URI;
 @RequiredArgsConstructor
 @Log4j2
 public class ClearingHouseService {
+
+    /**
+     * The clearing house API path for creating processes.
+     */
+    @Value("${clearing.house.path.process:}")
+    private String processPath;
+
+    /**
+     * The clearing house API path for logging messages.
+     */
+    @Value("${clearing.house.path.log:}")
+    private String logPath;
 
     /**
      * Service for configuring policy settings.
@@ -47,6 +69,15 @@ public class ClearingHouseService {
      */
     private final @NonNull LogMessageService logMessageSvc;
 
+    /**
+     * Service for ids request messages.
+     */
+    private final @NonNull ProcessCreationRequestService requestService;
+
+    /**
+     * Object mapper for mapping to JSON.
+     */
+    private final ObjectMapper objectMapper = new ObjectMapper();
 
     /**
      * Send contract agreement to clearing house.
@@ -82,6 +113,63 @@ public class ClearingHouseService {
         }
     }
 
+    /**
+     * Creates a process at the Clearing House using a contract agreement's UUID, where both the
+     * provider and the consumer of the agreement are registered as process owners. This is
+     * necessary so that both parties are able to log messages under the same process.
+     *
+     * @param agreement the contract agreement.
+     * @throws JsonProcessingException if parsing the process owners to JSON format fails.
+     */
+    public void createProcessAtClearingHouse(final ContractAgreement agreement)
+            throws JsonProcessingException {
+        if (isClearingHouseEnabled()) {
+            final var agreementId = UUIDUtils.uuidFromUri(agreement.getId());
+            final var url = buildProcessCreationUrl(agreementId);
+            final var payload = buildProcessCreationPayload(agreement.getProvider(),
+                    agreement.getConsumer());
+
+            final var response = requestService.send(new ProcessCreationMessageDesc(url),
+                    objectMapper.writeValueAsString(payload));
+
+            if (!requestService.isValidResponseType(response)) {
+                throw new MessageResponseException("Received unexpected response message type from"
+                        + " the Clearing House.");
+            }
+        }
+    }
+
+    /**
+     * Builds the URL for creating a new process at the Clearing House. The given process ID is
+     * appended to the Clearing House URL and the path for process creation.
+     *
+     * @param id the process ID.
+     * @return the URL.
+     */
+    private URI buildProcessCreationUrl(final UUID id) {
+        final var clearingHouse = connectorConfig.getClearingHouse();
+        final var uriBuilder = UriComponentsBuilder
+                .fromHttpUrl(clearingHouse.toString());
+        uriBuilder.pathSegment(processPath, id.toString());
+        return uriBuilder.build().toUri();
+    }
+
+    /**
+     * Builds the payload for creating a new process at the Clearing House. Both the provider and
+     * the consumer of an agreement are added as the process owners, so that both may log using
+     * the same ID.
+     *
+     * @param provider the provider ID.
+     * @param consumer the consumer ID.
+     * @return the payload.
+     */
+    private JSONObject buildProcessCreationPayload(final URI provider, final URI consumer) {
+        final var list = Arrays.asList(provider.toString(), consumer.toString());
+        final var payload = new JSONObject();
+        payload.put("owners", list);
+        return payload;
+    }
+
     private boolean isClearingHouseEnabled() {
         return !connectorConfig.getClearingHouse().toString().isBlank();
     }
@@ -89,7 +177,7 @@ public class ClearingHouseService {
     private URI buildDestination(final URI agreementId) {
         final var clearingHouse = connectorConfig.getClearingHouse();
         final var uriBuilder = UriComponentsBuilder.fromHttpUrl(clearingHouse.toString());
-        uriBuilder.pathSegment(UUIDUtils.uuidFromUri(agreementId).toString());
+        uriBuilder.pathSegment(logPath, UUIDUtils.uuidFromUri(agreementId).toString());
         return uriBuilder.build().toUri();
     }
 }